condensed Worker specs
